NASAExplorationRoversControl.ExplorationRover (NASA Exploration Rovers Control v1.0.0)

Exploration Rover entity, it represents a real exploration rover that is exploring another planet.

It has three attributes:

position (required): A tuble with x and y axis.

direction (required): A direction according to the Wind Rose Compass (in English) represented by a single character string. Valid values are N, S, W, E.

commands: A list of commands to be performed by rover. Valid values for each command in the list are L (left), R (right) or M (move).

Examples

iex> %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "N"}
%NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "N"}

iex> %NASAExplorationRoversControl.ExplorationRover{position: {3,2}, direction: "S"}
%NASAExplorationRoversControl.ExplorationRover{position: {3,2}, direction: "S"}

iex> %NASAExplorationRoversControl.ExplorationRover{position: {6,6}, direction: "W"}
%NASAExplorationRoversControl.ExplorationRover{position: {6,6}, direction: "W"}

iex> %NASAExplorationRoversControl.ExplorationRover{position: {3,0}, direction: "E"}
%NASAExplorationRoversControl.ExplorationRover{position: {3,0}, direction: "E"}

Types

t()

Specs

t() :: %NASAExplorationRoversControl.ExplorationRover{
  commands: List.t(),
  direction: String.t(),
  position: Tuple.t()
}

Functions

change_direction(exploration_rover, new_direction)

Changes existent exploration rover direction.

Examples

iex> exploration_rover = %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "N"}
...> exploration_rover |> NASAExplorationRoversControl.ExplorationRover.change_direction("W")
{:ok, %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "W"}}

iex> exploration_rover = %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "S"}
...> exploration_rover |> NASAExplorationRoversControl.ExplorationRover.change_direction("E")
{:ok, %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "E"}}

iex> exploration_rover = %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "S"}
...> exploration_rover |> NASAExplorationRoversControl.ExplorationRover.change_direction("I")
{:error, "Invalid direction. Must be N,S,W or E."}

change_position(exploration_rover, new_position)

Changes existent exploration rover position.

Examples

iex> exploration_rover = %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "N"}
...> exploration_rover |> NASAExplorationRoversControl.ExplorationRover.change_position({3,3})
{:ok, %NASAExplorationRoversControl.ExplorationRover{position: {3,3}, direction: "N"}}

iex> exploration_rover = %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "S"}
...> exploration_rover |> NASAExplorationRoversControl.ExplorationRover.change_position({1,4})
{:ok, %NASAExplorationRoversControl.ExplorationRover{position: {1,4}, direction: "S"}}

iex> exploration_rover = %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "S"}
...> exploration_rover |> NASAExplorationRoversControl.ExplorationRover.change_position({-1,4})
{:error, "Invalid position. Coordinates must not be negative."}

iex> exploration_rover = %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "S"}
...> exploration_rover |> NASAExplorationRoversControl.ExplorationRover.change_position(123)
{:error, "Invalid position. Must be a tuple."}

clear_commands(exploration_rover)

Clears an existing exploration rover given commands.

Examples

iex> exploration_rover = %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "N", commands: ["L"]}
...> exploration_rover |> NASAExplorationRoversControl.ExplorationRover.clear_commands()
{:ok, %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "N", commands: []}}

iex> exploration_rover = %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "N", commands: ["L","M","R"]}
...> exploration_rover |> NASAExplorationRoversControl.ExplorationRover.clear_commands()
{:ok, %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "N", commands: []}}

give_commands(exploration_rover, commands)

Adds commands to an existing exploration rover. Each command will be validated to match a valid command that can be L (left), R (right) or M (move).

Examples

iex> exploration_rover = %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "N"}
...> exploration_rover |> NASAExplorationRoversControl.ExplorationRover.give_commands(["L"])
{:ok, %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "N", commands: ["L"]}}

iex> exploration_rover = %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "N"}
...> exploration_rover |> NASAExplorationRoversControl.ExplorationRover.give_commands(["L", "L", "R"])
{:ok, %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "N", commands: ["L", "L", "R"]}}

iex> exploration_rover = %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "N"}
...> exploration_rover |> NASAExplorationRoversControl.ExplorationRover.give_commands([])
{:error, "Commands list must not be empty."}

iex> exploration_rover = %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "N"}
...> exploration_rover |> NASAExplorationRoversControl.ExplorationRover.give_commands("string")
{:error, "Commands must be a list."}

iex> exploration_rover = %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "N"}
...> exploration_rover |> NASAExplorationRoversControl.ExplorationRover.give_commands(["I"])
{:error, "Command `I` is invalid. Must be L, R or M."}

iex> exploration_rover = %NASAExplorationRoversControl.ExplorationRover{position: {0,0}, direction: "N"}
...> exploration_rover |> NASAExplorationRoversControl.ExplorationRover.give_commands(["L","M","P"])
{:error, "Command `P` is invalid. Must be L, R or M."}

new(list)

Returns a new struct for given position and direction. It also validates position and direction fields and raise when something is not right.

Examples

iex> NASAExplorationRoversControl.ExplorationRover.new(position: {10,3}, direction: "W")
{:ok, %NASAExplorationRoversControl.ExplorationRover{position: {10,3}, direction: "W"}}

iex> NASAExplorationRoversControl.ExplorationRover.new(position: {3,5}, direction: "S")
{:ok, %NASAExplorationRoversControl.ExplorationRover{position: {3,5}, direction: "S"}}

iex> NASAExplorationRoversControl.ExplorationRover.new(position: {-1,0}, direction: "E")
{:error, "Invalid position. Coordinates must not be negative."}

iex> NASAExplorationRoversControl.ExplorationRover.new(position: {0,-1}, direction: "E")
{:error, "Invalid position. Coordinates must not be negative."}

iex> NASAExplorationRoversControl.ExplorationRover.new(position: "string", direction: "E")
{:error, "Invalid position. Must be a tuple."}

iex> NASAExplorationRoversControl.ExplorationRover.new(position: 123, direction: "N")
{:error, "Invalid position. Must be a tuple."}

iex> NASAExplorationRoversControl.ExplorationRover.new(position: [], direction: "N")
{:error, "Invalid position. Must be a tuple."}

iex> NASAExplorationRoversControl.ExplorationRover.new(position: %{}, direction: "N")
{:error, "Invalid position. Must be a tuple."}

iex> NASAExplorationRoversControl.ExplorationRover.new(position: {0,0}, direction: "invalid")
{:error, "Invalid direction. Must be N,S,W or E."}

iex> NASAExplorationRoversControl.ExplorationRover.new(position: {0,0}, direction: "I")
{:error, "Invalid direction. Must be N,S,W or E."}

iex> NASAExplorationRoversControl.ExplorationRover.new(position: {0,0}, direction: "P")
{:error, "Invalid direction. Must be N,S,W or E."}

iex> NASAExplorationRoversControl.ExplorationRover.new(position: {0,0}, direction: 123)
{:error, "Invalid direction. Must be N,S,W or E."}

iex> NASAExplorationRoversControl.ExplorationRover.new(position: {0,0}, direction: [])
{:error, "Invalid direction. Must be N,S,W or E."}

iex> NASAExplorationRoversControl.ExplorationRover.new(position: {0,0}, direction: {})
{:error, "Invalid direction. Must be N,S,W or E."}

iex> NASAExplorationRoversControl.ExplorationRover.new(position: {0,0}, direction: %{})
{:error, "Invalid direction. Must be N,S,W or E."}